requirements.txt
#Pillow
#lxml
#Twisted
#numba
#numpy
#pandas

Django==2.2.4
Twisted[tls,http2]
channels
ldap3
django_python3_ldap
django-bootstrap4
django_select2
django-cors-headers
whitenoise
fpdf
html5lib
lxml
markdown2
requests
openpyxl
httpie
cffi
Pillow
fs
PyDispatcher
pypyodbc
daphne
polib
libsass
pyexcel_odsr
unidecode
django-bulk_update
dj-database-url
dj-email-url
easy_thumbnails
django-mptt
django_polymorphic
django-mailer
django_redis
django-cache-url
django-allauth
git+https://github.com/Splawik/django-filer.git@feature/django2-support
Transcrypt
django-sql-explorer
pendulum
cefpython3
xonsh
hypercorn
netifaces
cython
numba
numpy
numexpr
pandas
dukpy
